скорректирован тест по отзывам
